+++ Source/Editor/ConfigureReinteropEditor.cs
@@ -226,8 +226,8 @@ namespace CesiumForUnity
             ionOverlay = asset.overlay;
             server = ionOverlay.ionServer;
 
-            CesiumGeoJsonDocumentRasterOverlay geoJsonDocOverlay = asset.geoJsonOverlay;
-            server = geoJsonDocOverlay.ionServer;
+            CesiumCartographicPolygon cartographicPolygonAsset = asset.cartographicPolygon;
+            server = cartographicPolygonAsset.ionServer;
 
             AssetTroubleshootingDetails assetDetails = troubleshootingWindow.assetDetails;
             assetDetails.assetExistsInUserAccount = true;